Update to close Issue #11
The error was to do with OS support.
Windows refs cell [A1] as 0,0 whereas linux refs it as 1,1.
Therefore an error was thrown on linux as 0,0 does not exist. An offset
depending on OS.name has been factored into xlsLogging to solve the
issue. xlsLogging4Vars and 6Vars to be added into one class in the next
commit.
Issue #11 can now therefore be closed.